[fix] categoria, sub categoria y service
se solucionaron errores en endpoint de delete, getOne y relación entre Servicio y subcageboria

+++ src/category/service/services/categoryService.service.ts
@@ -6,6 +6,10 @@ import {CategoryServicesEntity} from "@/category/entities/services/categoryServi
 import {Repository} from "typeorm";
 import {InjectRepository} from "@nestjs/typeorm";
 import {CategoryServiceMapper} from "@/category/mapper/services/categoryService.mapper";
+import {SubCategoryService} from "@/category/service/subCategory.service";
+import {SubCategoryMapper} from "@/category/mapper/subCategory.mapper";
+import {SubCategoryEntity} from "@/category/entities/subCategory.entity";
+import {UpdateCategoryServiceDto} from "@/category/dto/services/updateCategoryService.dto";
 
 
 @Injectable()
@@ -14,20 +18,35 @@ export class CategoryServicesService {
   constructor(
       @InjectRepository(CategoryServicesEntity)
       private readonly categoryServicesRepository: Repository<CategoryServicesEntity>,
+      private readonly subCategoryService : SubCategoryService
   ) {}
 
   async create (createCategoryServicesDto : CreateCategoryServicesDto):Promise<ReadCategoryServicesDto>{
+    const subCategoryDto = this.subCategoryService.findOne(
+        new ValidID(createCategoryServicesDto.fkSubCategory)
+    )
+
+    if(!subCategoryDto) throw new HttpException(`subCategory with ID ${createCategoryServicesDto.fkSubCategory} not found`, HttpStatus.NOT_FOUND);
+
+    let entitySub = new SubCategoryEntity();
+    entitySub.pkSubCategory = createCategoryServicesDto.fkSubCategory;
+
+    let newEntity = this.categoryServicesRepository.create(createCategoryServicesDto);
+    newEntity.subCategory = entitySub;
+
+
     return CategoryServiceMapper.entityToReadServiceDto(
         await this.categoryServicesRepository.save(
-            this.categoryServicesRepository.create(createCategoryServicesDto)
+            newEntity
         )
     )
   }
 
   async findOne (validId : ValidID): Promise<ReadCategoryServicesDto>{
-    const entity = await this.categoryServicesRepository.findOneBy({
-      pkService : validId.id
-    })
+    const entity = await this.categoryServicesRepository.findOne({
+       where: {pkService : validId.id},
+       relations : ['subCategory','addons']
+    },)
     if(!entity){throw new HttpException(`CategoryServices with ID ${validId.id} not found`, HttpStatus.NOT_FOUND);}
 
     return CategoryServiceMapper.entityToReadServiceDto(entity);
@@ -42,10 +61,12 @@ export class CategoryServicesService {
 
 
   async remove (validID : ValidID): Promise<{ message: string; status: HttpStatus }> {
-    const found = await this.findOne(validID);
+    const found = await this.categoryServicesRepository.findOneBy({pkService:validID.id})
+
+    if(!found) return { message: "CategoryServices Not found", status: HttpStatus.NOT_FOUND }
 
     const responseDeleted = await this.categoryServicesRepository.remove(
-        this.categoryServicesRepository.create(found)
+        found
     );
 
     if(!responseDeleted) return { message: "CategoryServices Deleted", status: HttpStatus.NOT_MODIFIED }
@@ -57,15 +78,22 @@ export class CategoryServicesService {
   }
 
 
-  async update (readCategoryServicesDto:ReadCategoryServicesDto): Promise< { message: string; status: HttpStatus }> {
-    const entity = await this.categoryServicesRepository.findOneBy({
-      pkService : readCategoryServicesDto.pkService
+  async update (readCategoryServicesDto:UpdateCategoryServiceDto): Promise< { message: string; status: HttpStatus }> {
+    const entity = await this.categoryServicesRepository.findOne({
+      where: {pkService : readCategoryServicesDto.pkService}
     })
+    console.log(entity)
     if(!entity){throw new HttpException(`CategoryServices with ID ${readCategoryServicesDto.pkService} not found`, HttpStatus.NOT_FOUND);}
 
-    const update = await this.categoryServicesRepository.merge(
-        CategoryServiceMapper.readCategoryServicesDtoToEntity(readCategoryServicesDto)
-    )
+    let subCateEntity = new SubCategoryEntity();
+    subCateEntity.pkSubCategory = readCategoryServicesDto.fkSubCategory;
+
+    const merge = await this.categoryServicesRepository.merge(
+        entity,readCategoryServicesDto
+    );
+    entity.subCategory = subCateEntity;
+    
+const update = this.categoryServicesRepository.save(entity);
 
     if(!update)return { message: "CategoryServices Non Updated", status: HttpStatus.NOT_MODIFIED }
